+++ packages/core/src/components/cart/CartToggle/CartToggle.tsx
@@ -1,4 +1,4 @@
-import { ButtonIcon } from 'src/components/ui/Button'
+import Button from 'src/components/ui/Button'
 import Icon from 'src/components/ui/Icon'
 import { useCartToggleButton } from 'src/sdk/cart/useCartToggleButton'
 
@@ -6,8 +6,10 @@ function CartToggle() {
   const btnProps = useCartToggleButton()
 
   return (
-    <ButtonIcon
-      data-fs-button-cart="true"
+    <Button
+      variant="tertiary"
+      data-fs-button-icon
+      data-fs-button-cart
       aria-label={`Cart with ${btnProps['data-items']} items`}
       icon={<Icon name="ShoppingCart" width={32} height={32} />}
       {...btnProps}