+++ packages/core/src/components/common/Navbar/Navbar.tsx
@@ -3,8 +3,7 @@ import { Suspense, useRef, useState } from 'react'
 
 import CartToggle from 'src/components/cart/CartToggle'
 import SearchInput from 'src/components/search/SearchInput'
-import {
-  ButtonIcon,
+import Button, {
   ButtonSignIn,
   ButtonSignInFallback,
 } from 'src/components/ui/Button'
@@ -37,7 +36,9 @@ function Navbar() {
         <section data-fs-navbar-row>
           {!searchExpanded && (
             <>
-              <ButtonIcon
+              <Button
+                variant="tertiary"
+                data-fs-button-icon
                 data-fs-navbar-button-menu
                 aria-label="Open Menu"
                 icon={<Icon name="List" width={32} height={32} />}
@@ -59,7 +60,9 @@ function Navbar() {
             data-fs-navbar-search-expanded={searchExpanded}
           >
             {searchExpanded && (
-              <ButtonIcon
+              <Button
+                variant="tertiary"
+                data-fs-button-icon
                 data-fs-button-collapse
                 aria-label="Collapse search bar"
                 icon={<Icon name="CaretLeft" width={32} height={32} />}