+++ packages/core/src/components/common/Navbar/NavbarSlider.tsx
@@ -1,7 +1,6 @@
 import { Suspense } from 'react'
 
-import {
-  ButtonIcon,
+import Button, {
   ButtonSignIn,
   ButtonSignInFallback,
 } from 'src/components/ui/Button'
@@ -41,7 +40,9 @@ function NavbarSlider() {
           <Logo />
         </Link>
 
-        <ButtonIcon
+        <Button
+          variant="tertiary"
+          data-fs-button-icon
           data-fs-navbar-slider-button
           aria-label="Close Menu"
           icon={<Icon name="X" width={32} height={32} />}