+++ packages/core/src/components/ui/Button/index.ts
@@ -1,6 +1,5 @@
 export { default } from './Button'
 export type { ButtonProps } from './Button'
-export { default as ButtonIcon } from './ButtonIcon'
 export { default as ButtonBuy } from './ButtonBuy'
 export { default as ButtonLink } from './ButtonLink'
 export { default as ButtonSignIn } from './ButtonLink/ButtonSignIn'